+++ cmd/workspace/data-sources/data-sources.go
@@ -28,8 +28,8 @@ func New() *cobra.Command {
   grep to search the response from this API for the name of your SQL warehouse
   as it appears in Databricks SQL.
 
-  **Note**: A new version of the Databricks SQL API is now available. [Learn
-  more]
+  **Warning**: This API is deprecated. Please see the latest version of the
+  Databricks SQL API. [Learn more]
 
   [Learn more]: https://docs.databricks.com/en/sql/dbsql-api-latest.html`,
 		GroupID: "sql",
@@ -66,8 +66,8 @@ func newList() *cobra.Command {
   fields that appear in this API response are enumerated for clarity. However,
   you need only a SQL warehouse's id to create new queries against it.
 
-  **Note**: A new version of the Databricks SQL API is now available. Please use
-  :method:warehouses/list instead. [Learn more]
+  **Warning**: This API is deprecated. Please use :method:warehouses/list
+  instead. [Learn more]
 
   [Learn more]: https://docs.databricks.com/en/sql/dbsql-api-latest.html`